# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_uthmanduro_BracketForge_internal_model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  github_com_uthmanduro_BracketForge_internal_model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  github_com_uthmanduro_BracketForge_internal_model.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - organizer
        - viewer
        type: string
    required:
    - email
    - password
    - role
    type: object
  github_com_uthmanduro_BracketForge_internal_model.Role:
    enum:
    - admin
    - organizer
    - guest
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleOrganizer
    - RoleGuest
...
  github_com_uthmanduro_BracketForge_internal_model.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  github_com_uthmanduro_BracketForge_internal_model.User:
    properties:
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: string
      organization_id:
        type: string
      role:
        $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Role'
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bracketforge.io
    name: BracketForge Support
  description: Racket sports tournament management engine. Supports single elimination,
    round robin, and group + knockout formats with set-based scoring, three-way tiebreakers,
    BYE handling, and result correction.
  termsOfService: http://swagger.io/terms/
  title: BracketForge API
  version: 1.0.0
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticates with email and password. Returns a JWT token to use
        in the Authorization header.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      summary: Login
      tags:
      - auth
  /users/me:
    get:
      description: Returns the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /users/register:
    post:
      consumes:
      - application/json
      description: Creates a new user within the authenticated organisation. Admin
        only.
      parameters:
      - description: User details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"